+++ src/calendar.js
@@ -665,6 +665,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
         _element_View_FullYear_TitleBar = null,
         _element_View_FullYear_Contents = null,
         _element_View_FullYear_CurrentYear = null,
+        _element_View_FullYear_Contents_CurrentMonthTop = 0,
 
         // Variables: View - All Events
         _element_View_AllEvents = null,
@@ -4104,11 +4105,19 @@ function calendarJs( elementOrId, options, searchOptions ) {
         }
     }
 
-    function buildFullYearMonths() {
+    function buildFullYearMonths( fromOpen ) {
         _element_View_FullYear_Contents.innerHTML = _string.empty;
 
         for ( var monthIndex = 0; monthIndex < 12; monthIndex++ ) {
             buildFullYearViewMonth( monthIndex );
+        }
+
+        if ( fromOpen ) {
+            if ( _element_View_FullYear_Contents_CurrentMonthTop > 0 ) {
+                _element_View_FullYear_Contents.scrollTop = _element_View_FullYear_Contents_CurrentMonthTop - ( _options.spacing * 4 );
+            } else {
+                _element_View_FullYear_Contents.scrollTop = 0;
+            }
         }
     }
 
@@ -4158,9 +4167,15 @@ function calendarJs( elementOrId, options, searchOptions ) {
             buildFullYearViewMonthPreviousMonthDays( startDay, monthDayId, monthIndex );
         }
 
-        var lastFilledDay = buildFullYearViewMonthDays( startDay, monthDayId, monthIndex );
+        var currentMonthFillDetails = buildFullYearViewMonthDays( startDay, monthDayId, monthIndex ),
+            isCurrentMonth = currentMonthFillDetails[ 0 ],
+            lastFilledDay = currentMonthFillDetails[ 1 ];
 
         buildFullYearViewMonthNextMonthDays( lastFilledDay, monthDayId, monthIndex );
+
+        if ( isCurrentMonth ) {
+            _element_View_FullYear_Contents_CurrentMonthTop = yearMonth.offsetTop;
+        }
     }
 
     function buildFullYearViewMonthPreviousMonthDays( startDay, monthDayId, monthIndex ) {
@@ -4200,19 +4215,22 @@ function calendarJs( elementOrId, options, searchOptions ) {
 
     function buildFullYearViewMonthDays( startDay, monthDayId, monthIndex ) {
         var elementDayNumber = 0,
-            totalDaysInMonth = getTotalDaysInMonth( _element_View_FullYear_CurrentYear, monthIndex );
+            totalDaysInMonth = getTotalDaysInMonth( _element_View_FullYear_CurrentYear, monthIndex ),
+            isCurrentMonth = false;
 
         for ( var day = 0; day < totalDaysInMonth; day++ ) {
             elementDayNumber = startDay + day;
             
-            buildFullYearViewMonthDay( monthDayId, monthIndex, day, elementDayNumber );
+            isCurrentMonth = buildFullYearViewMonthDay( monthDayId, monthIndex, day, elementDayNumber );
         }
 
-        return elementDayNumber;
+        return [ isCurrentMonth, elementDayNumber ];
     }
 
     function buildFullYearViewMonthDay( monthDayId, monthIndex, day, elementDayNumber ) {
-        var currentMonthDayFullDayElement = getElementByID( monthDayId + elementDayNumber );
+        var currentMonthDayFullDayElement = getElementByID( monthDayId + elementDayNumber ),
+            isCurrentMonth = false;
+
         if ( currentMonthDayFullDayElement !== null ) {
             var currentMonthDayDate = new Date( _element_View_FullYear_CurrentYear, monthIndex, day + 1 );
 
@@ -4224,9 +4242,11 @@ function calendarJs( elementOrId, options, searchOptions ) {
                 showDayContextMenu( e, currentMonthDayDate );
             };
 
-            buildFullYearViewMonthDayClasses( currentMonthDayFullDayElement, currentMonthDayDate, true );
+            isCurrentMonth = buildFullYearViewMonthDayClasses( currentMonthDayFullDayElement, currentMonthDayDate, true );
             buildDayDisplay( currentMonthDayFullDayElement, currentMonthDayDate );
         }
+
+        return isCurrentMonth;
     }
 
     function buildFullYearViewMonthNextMonthDays( lastDayFilled, monthDayId, monthIndex  ) {
@@ -4263,9 +4283,10 @@ function calendarJs( elementOrId, options, searchOptions ) {
     }
 
     function buildFullYearViewMonthDayClasses( element, date, showTodayCss ) {
-        showTodayCss = isDefined( showTodayCss ) ? showTodayCss : false;
+        var formattedDate = toStorageFormattedDate( date ),
+            isCurrentMonth = false;
 
-        var formattedDate = toStorageFormattedDate( date );
+        showTodayCss = isDefined( showTodayCss ) ? showTodayCss : false;
 
         if ( isWeekendDay( date ) ) {
             element.className += " weekend-day";
@@ -4287,21 +4308,21 @@ function calendarJs( elementOrId, options, searchOptions ) {
 
         if ( showTodayCss && isDateToday( date ) ) {
             element.className += " cell-today";
+            isCurrentMonth = true;
         }
+
+        return isCurrentMonth;
     }
 
     function showFullYearView( year, fromOpen ) {
         fromOpen = isDefined( fromOpen ) ? fromOpen : false;
         _element_View_FullYear_CurrentYear = isDefined( year ) ? year : _currentDate.getFullYear();
 
-        if ( fromOpen ) {
-            _element_View_FullYear_Contents.scrollTop = 0;
-        }
-
-        buildFullYearMonths();
-        showView( _element_View_FullYear );
-
+        _element_View_FullYear_Contents_CurrentMonthTop = 0;
         _element_View_FullYear_TitleBar.innerText = _element_View_FullYear_CurrentYear;
+
+        buildFullYearMonths( fromOpen );
+        showView( _element_View_FullYear );
     }
 
     function updateFullYearView() {